# app/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "System Rejestracji Pacjenta Stomatologicznego API",
      "description": "API dla systemu rejestracji pacjenta stomatologicznego",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/api",
        "description": "Development server"
      }
    ],
    "components": {
      "securitySchemes": {
        "jwt": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "paths": {
...
      "/auth/login": {
        "post": {
          "tags": ["Autoryzacja"],
          "summary": "Logowanie użytkownika",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": {"type": "string"},
                    "haslo": {"type": "string"},
                    "rola": {"type": "string", "enum": ["pacjent", "lekarz", "admin"]}
                  },
                  "required": ["email", "haslo", "rola"]
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Zalogowano pomyślnie"
            },
            "401": {
              "description": "Nieprawidłowy email lub hasło"
            }
          }
        }
      },
      "/wizyty/terminy": {
        "get": {
          "tags": ["Wizyty"],
          "summary": "Pobieranie dostępnych terminów",
          "parameters": [
            {
              "name": "lekarz_id",
              "in": "query",
              "schema": {
                "type": "integer"
              }
            },
            {
              "name": "data_od",
              "in": "query",
              "schema": {
                "type": "string",
                "format": "date"
              }
            },
            {
              "name": "data_do",
              "in": "query",
              "schema": {
                "type": "string",
                "format": "date"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Lista dostępnych terminów"
            }
          }
        }
      },
...
      "/lekarze": {
        "get": {
          "tags": ["Lekarze"],
          "summary": "Pobieranie listy lekarzy",
          "parameters": [
            {
              "name": "specjalizacja",
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "aktywni",
              "in": "query",
              "schema": {
                "type": "boolean"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Lista lekarzy"
            }
          }
        }
      },
...
    }
  }